Create Best of 5 (Easy, Medium and Hard)
Its a Simple game of Tick Tak Toe but against computer and User gets to Select the Difficulty of the game which are classified as Easy, Medium and Hard. The game is designed in a way that it is Best of 5 so If the player Wins 3 times then he shall be declared as the winner while if he looses then he shall Accept that he was Defeated.